pyproject.toml
[project]
name = "bcm2711-deskpet-panel"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "numpy>=2.4.0",
    "pillow>=12.1.0",
    "psutil>=7.2.1",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
    "rpi-gpio>=0.7.1",
    "spidev>=3.8",
]
